# server/swaggerDoc.json
{
  "swagger": "2.0",
  "info": {
    "description": " This is the api documentation for the Monash Tutor Allocation System",
    "version": "1.0.0",
    "title": "Monash Tutor Allocation System"
  },
...
  "schemes": ["http"],
  "paths": {
...
    "/roles/unit/{unitId}": {
      "get": {
        "tags": ["Role"],
        "summary": "Gets a unit's roles",
        "operationId": "getRolesByUnit",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "unitId",
            "description": "Unit Id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": { "type": "string", "format": "uuid" },
                  "title": { "type": "string", "enum": ["TA", "Lecturer"] },
                  "unitId": { "type": "string", "format": "uuid" },
                  "staffId": { "type": "string", "format": "uuid" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Role"],
        "summary": "Creates a role record for the unit",
        "operationId": "createRole",
        "consumes": ["application.json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "newRecord",
            "description": "new role object",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "unitId",
            "description": "Unit Id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "id": { "type": "string", "format": "uuid" },
                "title": { "type": "string", "enum": ["TA", "Lecturer"] },
                "unitId": { "type": "string", "format": "uuid" },
                "staffId": { "type": "string", "format": "uuid" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Role"],
        "summary": "Updates a role record for the unit",
        "operationId": "updateRole",
        "consumes": ["application.json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "changedRecord",
            "description": "updated role object",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "unitId",
            "description": "Unit Id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "id": { "type": "string", "format": "uuid" },
                "title": { "type": "string", "enum": ["TA", "Lecturer"] },
                "unitId": { "type": "string", "format": "uuid" },
                "staffId": { "type": "string", "format": "uuid" }
              }
            }
          }
        }
      }
    },
    "/roles/unit/{unitId}/me": {
      "get": {
        "tags": ["Role"],
        "summary": "Get role of the user for a unit",
        "operationId": "getMyRoleByUnit",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "unitId",
            "description": "Unit Id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "id": { "type": "string", "format": "uuid" },
                "title": { "type": "string", "enum": ["TA", "Lecturer"] },
                "unitId": { "type": "string", "format": "uuid" },
                "staffId": { "type": "string", "format": "uuid" }
              }
            }
          }
        }
      }
    },
    "/roles/unit/{unitId}/{roleId}": {
      "delete": {
        "tags": ["Role"],
        "summary": "Deletes a role",
        "operationId": "deleteRole",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "unitId",
            "description": "Unit Id",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "roleId",
            "description": "Role Id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "id": { "type": "string", "format": "uuid" },
                "title": { "type": "string", "enum": ["TA", "Lecturer"] },
                "unitId": { "type": "string", "format": "uuid" },
                "staffId": { "type": "string", "format": "uuid" }
              }
            }
          }
        }
      }
    },
...
  },
...
}